refactor: dedupe helpers and simplify hot paths
- Fix MCP IPC marker scanning in adapter stdout buffer
- Remove dead lastResizes map and one-off input-ready wrapper
- Share levenshtein, readProxyResponse, isNodeboxRuntime across runtime
- Use accChunks join and compact tool-result JSON in agent turn
- Reuse formatBytes/errorMessage in orchestrator
